fix: make kuzu dependency optional behind kuzu-backend feature flag
cargo install --git without cmake now works by default.
Users needing the kuzu graph backend must pass --features kuzu-backend
and have cmake installed. Also removes stale flat-file module duplicates
(memory.rs and recipe.rs) that were shadowing their directory equivalents.
