+++ crates/amplihack-cli/Cargo.toml
@@ -5,6 +5,9 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 
+[features]
+kuzu-backend = ["dep:kuzu"]
+
 [dependencies]
 amplihack-types = { workspace = true }
 amplihack-state = { workspace = true }
@@ -22,7 +25,7 @@ tempfile = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 ureq = { version = "2", features = ["json"] }
-kuzu = "0.11.3"
+kuzu = { version = "0.11.3", optional = true }
 rusqlite = { version = "0.38.0", features = ["bundled"] }
 
 [dev-dependencies]
